update monaco editor schema when endpoint url change
